+++ client/src/main/java/replicant/MessageResponse.java
@@ -17,7 +17,6 @@ import static org.realityforge.braincheck.Guards.*;
 /**
  * A simple class encapsulating the process of loading data from a json change set.
  */
-@SuppressFBWarnings( value = { "EQ_COMPARETO_USE_OBJECT_EQUALS" }, justification = "Equals is not used and implementing it would add code that GWT2.x could not optimize away" )
 final class MessageResponse
 {
   private final int _schemaId;