+++ client/src/main/java/replicant/messages/AbstractMessage.java
@@ -9,7 +9,6 @@ import jsinterop.annotations.JsType;
  * The abstract message that messages conform to.
  */
 @SuppressWarnings( "NotNullFieldNotInitialized" )
-@SuppressFBWarnings( "EI_EXPOSE_REP" )
 @JsType( isNative = true, namespace = JsPackage.GLOBAL, name = "Object" )
 public abstract class AbstractMessage
 {