+++ client/src/main/java/replicant/messages/ErrorMessage.java
@@ -9,7 +9,6 @@ import replicant.shared.Messages;
 /**
  * The message that represents a set of changes to subscriptions and entities that should be applied atomically.
  */
-@SuppressFBWarnings( "EI_EXPOSE_REP" )
 @JsType( isNative = true, namespace = JsPackage.GLOBAL, name = "Object" )
 public class ErrorMessage
   extends ServerToClientMessage