+++ client/src/main/java/replicant/messages/UseCacheMessage.java
@@ -12,7 +12,6 @@ import replicant.shared.Messages;
  * The message that represents a set of changes to subscriptions and entities that should be applied atomically.
  */
 @SuppressWarnings( "NotNullFieldNotInitialized" )
-@SuppressFBWarnings( "EI_EXPOSE_REP" )
 @JsType( isNative = true, namespace = JsPackage.GLOBAL, name = "Object" )
 public class UseCacheMessage
   extends ServerToClientMessage