+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v6.161](https://github.com/replicant4j/replicant/tree/v6.161) (2025-10-15) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.160...v6.161)
 
 Changes in this release: